Code refactor after PR
